+++ CHANGES.md
@@ -1,2 +1,4 @@
-- Added Vault of the Incarnates tier set tokens
-- Handle need-before-green with masterlooter like personal loot for now
+- Update ToC version for patch 10.1
+- Add support for omni tokens
+- Add Aberrus tier set tokens
+- Fix loot method detection for raids